main.py refactor
The idea here is to focus main.py into mainly having the event handler registration, and any additional, re-usable logic can be offloaded to utils.py.

The load / unload commands were changed to be more self contained in terms of logic and to return a boolean rather than an exit-code like number.